+++ lib/connection.js
@@ -156,7 +156,7 @@ class Connection extends EventEmitter {
   }
 
   promise(promiseImpl) {
-    const PromiseConnection = require('../promise').PromiseConnection;
+    const PromiseConnection = require('./promise_connection.js');
     return new PromiseConnection(this, promiseImpl);
   }
