+++ lib/pool.js
@@ -33,7 +33,7 @@ class Pool extends EventEmitter {
   }
 
   promise(promiseImpl) {
-    const PromisePool = require('../promise').PromisePool;
+    const PromisePool = require('./promise_pool.js');
     return new PromisePool(this, promiseImpl);
   }
