+++ lib/pool_connection.js
@@ -30,7 +30,7 @@ class PoolConnection extends Connection {
   }
 
   promise(promiseImpl) {
-    const PromisePoolConnection = require('../promise').PromisePoolConnection;
+    const PromisePoolConnection = require('./promise_pool_connection.js');
     return new PromisePoolConnection(this, promiseImpl);
   }
