+++ promise.js
@@ -3,257 +3,17 @@
 const SqlString = require('sqlstring');
 const EventEmitter = require('events').EventEmitter;
 const parserCache = require('./lib/parsers/parser_cache.js');
-const PoolConnection = require('./lib/pool_connection.js');
 const Pool = require('./lib/pool.js');
 const PoolCluster = require('./lib/pool_cluster.js');
 const createConnection = require('./lib/create_connection.js');
 const Connection = require('./lib/connection.js');
 const createPool = require('./lib/create_pool.js');
 const createPoolCluster = require('./lib/create_pool_cluster.js');
-
-function makeDoneCb(resolve, reject, localErr) {
-  return function (err, rows, fields) {
-    if (err) {
-      localErr.message = err.message;
-      localErr.code = err.code;
-      localErr.errno = err.errno;
-      localErr.sql = err.sql;
-      localErr.sqlState = err.sqlState;
-      localErr.sqlMessage = err.sqlMessage;
-      reject(localErr);
-    } else {
-      resolve([rows, fields]);
-    }
-  };
-}
-
-function inheritEvents(source, target, events) {
-  const listeners = {};
-  target
-    .on('newListener', eventName => {
-      if (events.indexOf(eventName) >= 0 && !target.listenerCount(eventName)) {
-        source.on(
-          eventName,
-          (listeners[eventName] = function () {
-            const args = [].slice.call(arguments);
-            args.unshift(eventName);
-
-            target.emit.apply(target, args);
-          })
-        );
-      }
-    })
-    .on('removeListener', eventName => {
-      if (events.indexOf(eventName) >= 0 && !target.listenerCount(eventName)) {
-        source.removeListener(eventName, listeners[eventName]);
-        delete listeners[eventName];
-      }
-    });
-}
-
-class PromisePreparedStatementInfo {
-  constructor(statement, promiseImpl) {
-    this.statement = statement;
-    this.Promise = promiseImpl;
-  }
-
-  execute(parameters) {
-    const s = this.statement;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      if (parameters) {
-        s.execute(parameters, done);
-      } else {
-        s.execute(done);
-      }
-    });
-  }
-
-  close() {
-    return new this.Promise(resolve => {
-      this.statement.close();
-      resolve();
-    });
-  }
-}
-
-class PromiseConnection extends EventEmitter {
-  constructor(connection, promiseImpl) {
-    super();
-    this.connection = connection;
-    this.Promise = promiseImpl || Promise;
-    inheritEvents(connection, this, [
-      'error',
-      'drain',
-      'connect',
-      'end',
-      'enqueue'
-    ]);
-  }
-
-  release() {
-    this.connection.release();
-  }
-
-  query(query, params) {
-    const c = this.connection;
-    const localErr = new Error();
-    if (typeof params === 'function') {
-      throw new Error(
-        'Callback function is not available with promise clients.'
-      );
-    }
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      if (params !== undefined) {
-        c.query(query, params, done);
-      } else {
-        c.query(query, done);
-      }
-    });
-  }
-
-  execute(query, params) {
-    const c = this.connection;
-    const localErr = new Error();
-    if (typeof params === 'function') {
-      throw new Error(
-        'Callback function is not available with promise clients.'
-      );
-    }
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      if (params !== undefined) {
-        c.execute(query, params, done);
-      } else {
-        c.execute(query, done);
-      }
-    });
-  }
-
-  end() {
-    return new this.Promise(resolve => {
-      this.connection.end(resolve);
-    });
-  }
-
-  beginTransaction() {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      c.beginTransaction(done);
-    });
-  }
-
-  commit() {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      c.commit(done);
-    });
-  }
-
-  rollback() {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      c.rollback(done);
-    });
-  }
-
-  ping() {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      c.ping(err => {
-        if (err) {
-          localErr.message = err.message;
-          localErr.code = err.code;
-          localErr.errno = err.errno;
-          localErr.sqlState = err.sqlState;
-          localErr.sqlMessage = err.sqlMessage;
-          reject(localErr);
-        } else {
-          resolve(true);
-        }
-      });
-    });
-  }
-
-  connect() {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      c.connect((err, param) => {
-        if (err) {
-          localErr.message = err.message;
-          localErr.code = err.code;
-          localErr.errno = err.errno;
-          localErr.sqlState = err.sqlState;
-          localErr.sqlMessage = err.sqlMessage;
-          reject(localErr);
-        } else {
-          resolve(param);
-        }
-      });
-    });
-  }
-
-  prepare(options) {
-    const c = this.connection;
-    const promiseImpl = this.Promise;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      c.prepare(options, (err, statement) => {
-        if (err) {
-          localErr.message = err.message;
-          localErr.code = err.code;
-          localErr.errno = err.errno;
-          localErr.sqlState = err.sqlState;
-          localErr.sqlMessage = err.sqlMessage;
-          reject(localErr);
-        } else {
-          const wrappedStatement = new PromisePreparedStatementInfo(
-            statement,
-            promiseImpl
-          );
-          resolve(wrappedStatement);
-        }
-      });
-    });
-  }
-
-  changeUser(options) {
-    const c = this.connection;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      c.changeUser(options, err => {
-        if (err) {
-          localErr.message = err.message;
-          localErr.code = err.code;
-          localErr.errno = err.errno;
-          localErr.sqlState = err.sqlState;
-          localErr.sqlMessage = err.sqlMessage;
-          reject(localErr);
-        } else {
-          resolve();
-        }
-      });
-    });
-  }
-
-  get config() {
-    return this.connection.config;
-  }
-
-  get threadId() {
-    return this.connection.threadId;
-  }
-}
+const PromiseConnection = require('./lib/promise_connection.js');
+const PromisePool = require('./lib/promise_pool.js');
+const makeDoneCb = require('./lib/make_done_cb.js');
+const PromisePoolConnection = require('./lib/promise_pool_connection.js');
+const inheritEvents = require('./lib/inherit_events.js');
 
 function createConnectionPromise(opts) {
   const coreConnection = createConnection(opts);
@@ -319,100 +79,6 @@ function createConnectionPromise(opts) {
   'resume',
   'unprepare'
 ]);
-
-class PromisePoolConnection extends PromiseConnection {
-  constructor(connection, promiseImpl) {
-    super(connection, promiseImpl);
-  }
-
-  destroy() {
-    return PoolConnection.prototype.destroy.apply(
-      this.connection,
-      arguments
-    );
-  }
-}
-
-class PromisePool extends EventEmitter {
-  constructor(pool, thePromise) {
-    super();
-    this.pool = pool;
-    this.Promise = thePromise || Promise;
-    inheritEvents(pool, this, ['acquire', 'connection', 'enqueue', 'release']);
-  }
-
-  getConnection() {
-    const corePool = this.pool;
-    return new this.Promise((resolve, reject) => {
-      corePool.getConnection((err, coreConnection) => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(new PromisePoolConnection(coreConnection, this.Promise));
-        }
-      });
-    });
-  }
-
-  releaseConnection(connection) {
-    if (connection instanceof PromisePoolConnection) connection.release();
-  }
-
-  query(sql, args) {
-    const corePool = this.pool;
-    const localErr = new Error();
-    if (typeof args === 'function') {
-      throw new Error(
-        'Callback function is not available with promise clients.'
-      );
-    }
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      if (args !== undefined) {
-        corePool.query(sql, args, done);
-      } else {
-        corePool.query(sql, done);
-      }
-    });
-  }
-
-  execute(sql, args) {
-    const corePool = this.pool;
-    const localErr = new Error();
-    if (typeof args === 'function') {
-      throw new Error(
-        'Callback function is not available with promise clients.'
-      );
-    }
-    return new this.Promise((resolve, reject) => {
-      const done = makeDoneCb(resolve, reject, localErr);
-      if (args) {
-        corePool.execute(sql, args, done);
-      } else {
-        corePool.execute(sql, done);
-      }
-    });
-  }
-
-  end() {
-    const corePool = this.pool;
-    const localErr = new Error();
-    return new this.Promise((resolve, reject) => {
-      corePool.end(err => {
-        if (err) {
-          localErr.message = err.message;
-          localErr.code = err.code;
-          localErr.errno = err.errno;
-          localErr.sqlState = err.sqlState;
-          localErr.sqlMessage = err.sqlMessage;
-          reject(localErr);
-        } else {
-          resolve();
-        }
-      });
-    });
-  }
-}
 
 function createPromisePool(opts) {
   const corePool = createPool(opts);
